isolate JAX/XLA into separate process to ensure GPU is usable later
